requirements.txt
dash
dash_bootstrap_components
dash_core_components
dash_html_components
# DeepExpressions
Pillow
plotly
